Move the last top level files inside gm/
PiperOrigin-RevId: 753134430
